chore: removed `http.ts` in favor of `client.ts`
