+++ store/client.ts
@@ -1,8 +1,27 @@
-import { Ast } from './http';
-
 const base = 'http://localhost:33333';
 
-type Response = {
+export type Table = { schema: string; table: string; alias: string };
+export type TableHint = {
+  schema: string;
+  table: string;
+  column: string;
+  parent: boolean;
+  pine: string;
+};
+
+export type Hints = { table: TableHint[] };
+// There are more operations. I'll add them as we need to handle them here
+export type Operation = { type: 'table' | 'delete' };
+
+export type Ast = {
+  hints: Hints;
+  'selected-tables': Table[];
+  joins: string[][];
+  context: string;
+  operation: Operation;
+};
+
+export type Response = {
   result: (string | number)[][];
   'connection-id': string;
   query: string;
@@ -12,7 +31,20 @@ type Response = {
 };
 
 export class HttpClient {
-  constructor() {}
+  constructor(private readonly onBuild?: (ast: Ast) => void) {}
+
+  public async get(path: string): Promise<Response | undefined> {
+    const res = await fetch(`${base}/api/v1/${path}`, {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    });
+    if (!res.ok) {
+      return;
+    }
+    return await res.json();
+  }
 
   private async post(path: string, body: object): Promise<Response | undefined> {
     const res = await fetch(`${base}/api/v1/${path}`, {
@@ -55,12 +87,12 @@ export class HttpClient {
   }
 
   public async build(expression: string): Promise<Response> {
-    // const x = this.cleanExpression(expression);
-    const x = expression;
+    const x = this.cleanExpression(expression);
     const response = await this.post('build', { expression: x });
     if (!response) {
       throw new Error('No response when trying to build');
     }
+    this.onBuild && (await this.onBuild(response.ast));
     return response;
   }
 
@@ -86,7 +118,7 @@ export class HttpClient {
     if (!response) {
       throw new Error('No response when trying to make child Expressions');
     }
-    // this.onBuild && (await this.onBuild(response.state));
+    this.onBuild && (await this.onBuild(response.ast));
     const expressions = response.ast.hints.table.filter(h => !h.parent).map(h => `${x} ${h.pine}`);
     return { expressions, ast: response.ast };
   }
